Correction after rebase + delete Suppliers of ConversionOption
